# lib/api-zod/src/generated/types/judgmentResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Prompt Engineering Game API
 * OpenAPI spec version: 0.1.0
 */
import type { Submission } from "./submission";

export interface JudgmentResult {
  roundId: number;
  submissions: Submission[];
  summary: string;
}