# lib/api-zod/src/generated/types/leaderboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Prompt Engineering Game API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from "./leaderboardEntry";

export interface Leaderboard {
  roomCode: string;
  players: LeaderboardEntry[];
}