# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Prompt Engineering Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  name: string;
  roomId: number;
  totalScore: number;
  joinedAt: Date;
}