# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Prompt Engineering Game API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomStatus } from "./roomStatus";

export interface Room {
  id: number;
  code: string;
  hostName: string;
  status: RoomStatus;
  createdAt: Date;
}