# lib/api-zod/src/generated/types/roomWithDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Prompt Engineering Game API
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from "./player";
import type { RoomWithDetailsStatus } from "./roomWithDetailsStatus";
import type { RoundWithSubmissions } from "./roundWithSubmissions";

export interface RoomWithDetails {
  id: number;
  code: string;
  hostName: string;
  status: RoomWithDetailsStatus;
  createdAt: Date;
  players: Player[];
  currentRound?: RoundWithSubmissions;
}