# lib/api-zod/src/generated/types/round.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Prompt Engineering Game API
 * OpenAPI spec version: 0.1.0
 */
import type { RoundStatus } from "./roundStatus";

export interface Round {
  id: number;
  roomId: number;
  roundNumber: number;
  category: string;
  prompt: string;
  status: RoundStatus;
  timeLimit: number;
  createdAt: Date;
}