# lib/api-zod/src/generated/types/roundWithSubmissions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Prompt Engineering Game API
 * OpenAPI spec version: 0.1.0
 */
import type { RoundWithSubmissionsStatus } from "./roundWithSubmissionsStatus";
import type { Submission } from "./submission";

export interface RoundWithSubmissions {
  id: number;
  roomId: number;
  roundNumber: number;
  category: string;
  prompt: string;
  status: RoundWithSubmissionsStatus;
  timeLimit: number;
  createdAt: Date;
  submissions: Submission[];
}